+++ pkg/analyzer/lib/src/dart/analysis/context_locator.dart
@@ -26,6 +26,7 @@ import 'package:analyzer/src/dart/analysis/performance_logger.dart'
 import 'package:analyzer/src/generated/sdk.dart' show DartSdkManager;
 import 'package:analyzer/src/task/options.dart';
 import 'package:analyzer/src/util/yaml.dart';
+import 'package:cli_util/cli_util.dart';
 import 'package:glob/glob.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart';
@@ -56,7 +57,9 @@ class ContextLocatorImpl implements ContextLocator {
       List<String> excludedPaths = const <String>[],
       String optionsFile,
       String packagesFile,
-      @required String sdkPath}) {
+      String sdkPath}) {
+    // TODO(scheglov) Remove this, and make `sdkPath` required.
+    sdkPath ??= getSdkPath();
     ArgumentError.checkNotNull(sdkPath, 'sdkPath');
 
     List<ContextRoot> roots = locateRoots(