+++ pkg/analyzer/CHANGELOG.md
@@ -4,6 +4,9 @@
   find SDK, for example using `package:cli_util`.
 * Fixed `LintDriver` for the new way to access `WorkspacePackage`.
 * Deprecated `Scope.lookup()`, use `Scope.lookup2()` instead.
+* Updated implementations of `ContextBuilder.createContext()` and
+  `ContextLocator.locateContexts` to use the current SDK if `sdkPath`
+  is not provided.
 
 ## 0.39.13
 * Added 'dart/sdk/build_sdk_summary.dart' with `buildSdkSummary`.